+++ spec/deimos_spec.rb
@@ -57,26 +57,13 @@ describe Deimos do
     expect(Deimos::VERSION).not_to be_nil
   end
 
-  specify 'configure' do
-    expect(Phobos).to receive(:configure).with(phobos_configuration)
-    allow(described_class).to receive(:ssl_var_contents) { |key| key }
-    described_class.configure do |config|
-      config.phobos_config_file = config_path
-      config.seed_broker = 'my_seed_broker.com'
-      config.ssl_enabled = true
-      config.ssl_ca_cert = 'my_ssl_ca_cert'
-      config.ssl_client_cert = 'my_ssl_client_cert'
-      config.ssl_client_cert_key = 'my_ssl_client_cert_key'
-    end
-  end
-
   it 'should error if required_acks is not all' do
     expect {
       described_class.configure do |config|
-        config.publish_backend = :db
+        config.producers.backend = :db
         config.phobos_config_file = File.join(File.dirname(__FILE__), 'phobos.bad_db.yml')
       end
-    }.to raise_error('Cannot set publish_backend to :db unless required_acks is set to ":all" in phobos.yml!')
+    }.to raise_error('Cannot set producers.backend to :db unless producers.required_acks is set to ":all"!')
   end
 
   describe '#start_db_backend!' do
@@ -94,7 +81,7 @@ describe Deimos do
         signal_handler
       end
       described_class.configure do |config|
-        config.publish_backend = :db
+        config.producers.backend = :db
       end
       described_class.start_db_backend!(thread_count: 2)
     end
@@ -102,7 +89,7 @@ describe Deimos do
     it 'should not start if backend is not db' do
       expect(Deimos::Utils::SignalHandler).not_to receive(:new)
       described_class.configure do |config|
-        config.publish_backend = :kafka
+        config.producers.backend = :kafka
       end
       expect { described_class.start_db_backend!(thread_count: 2) }.
         to raise_error('Publish backend is not set to :db, exiting')
@@ -111,7 +98,7 @@ describe Deimos do
     it 'should not start if thread_count is nil' do
       expect(Deimos::Utils::SignalHandler).not_to receive(:new)
       described_class.configure do |config|
-        config.publish_backend = :db
+        config.producers.backend = :db
       end
       expect { described_class.start_db_backend!(thread_count: nil) }.
         to raise_error('Thread count is not given or set to zero, exiting')
@@ -120,7 +107,7 @@ describe Deimos do
     it 'should not start if thread_count is 0' do
       expect(Deimos::Utils::SignalHandler).not_to receive(:new)
       described_class.configure do |config|
-        config.publish_backend = :db
+        config.producers.backend = :db
       end
       expect { described_class.start_db_backend!(thread_count: 0) }.
         to raise_error('Thread count is not given or set to zero, exiting')
@@ -128,11 +115,11 @@ describe Deimos do
 
     describe 'delivery configuration' do
       before(:each) do
-        described_class.config = nil
         allow(YAML).to receive(:load).and_return(phobos_configuration)
       end
 
       it 'should not raise an error with properly configured handlers' do
+        path = config_path # for scope issues in the block below
         # Add explicit consumers
         phobos_configuration['listeners'] << { 'handler' => 'ConsumerTest::MyConsumer',
                                                'delivery' => 'message' }
@@ -140,47 +127,39 @@ describe Deimos do
                                                'delivery' => 'batch' }
 
         expect {
-          described_class.configure { |c| c.phobos_config_file = config_path }
+          described_class.configure { |c| c.phobos_config_file = path }
         }.not_to raise_error
       end
 
       it 'should raise an error if BatchConsumers do not have inline_batch delivery' do
+        path = config_path # for scope issues in the block below
         phobos_configuration['listeners'] = [{ 'handler' => 'ConsumerTest::MyBatchConsumer',
                                                'delivery' => 'message' }]
 
         expect {
-          described_class.configure { |c| c.phobos_config_file = config_path }
+          described_class.configure { |c| c.phobos_config_file = path }
         }.to raise_error('BatchConsumer ConsumerTest::MyBatchConsumer must have delivery set to `inline_batch`')
       end
 
       it 'should raise an error if Consumers do not have message or batch delivery' do
+        path = config_path # for scope issues in the block below
         phobos_configuration['listeners'] = [{ 'handler' => 'ConsumerTest::MyConsumer',
                                                'delivery' => 'inline_batch' }]
 
         expect {
-          described_class.configure { |c| c.phobos_config_file = config_path }
+          described_class.configure { |c| c.phobos_config_file = path }
         }.to raise_error('Non-batch Consumer ConsumerTest::MyConsumer must have delivery set to `message` or `batch`')
       end
 
       it 'should treat nil as `batch`' do
+        path = config_path # for scope issues in the block below
         phobos_configuration['listeners'] = [{ 'handler' => 'ConsumerTest::MyConsumer' }]
 
         expect {
-          described_class.configure { |c| c.phobos_config_file = config_path }
+          described_class.configure { |c| c.phobos_config_file = path }
         }.not_to raise_error
       end
 
-      it 'should ignore non-Deimos listeners' do
-        consumer_class = Class.new { include Phobos::Handler }
-        stub_const('ConsumerTest::MyOtherConsumer', consumer_class)
-        phobos_configuration['listeners'] = [{ 'handler' => 'ConsumerTest::MyOtherConsumer',
-                                               'topic' => 'my_consume_topic',
-                                               'group_id' => 'my_group_id' }]
-
-        expect {
-          described_class.configure { |c| c.phobos_config_file = config_path }
-        }.not_to raise_error
-      end
     end
   end
 end