+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 
+# [1.4.0-beta7] - 2019-12-16
+- Clone loggers when assigning to multiple levels.
+
+# [1.4.0-beta6] - 2019-12-16
+- Added default for max_bytes_per_partition.
+
+# [1.4.0-beta4] - 2019-11-26
+- Added `define_settings` to define settings without invoking callbacks.
+
+# [1.4.0-beta2] - 2019-11-22
+- FIX: settings with default_proc were being called immediately
+  instead of being lazy-evaluated.
+
+# [1.4.0-beta1] - 2019-11-22
+- Complete revamp of configuration method.
+
 # [1.3.0-beta4] - 2019-12-02
 - Fixed bug where by running `rake deimos:start` without
   specifying a producer backend would crash.